// springflow-react/src/springflow/generated/dev/springflow/demo/employees/EmployeeEndpoint.ts
/* Generated by SpringFlow. DO NOT EDIT. */

import { SpringFlowClient } from "../../../../SpringFlowClient";
import type { Employee } from "./Employee";

export class EmployeeEndpoint {

    static async create(name: string, email: string, department: string): Promise<Employee> {
        return SpringFlowClient.get<Employee>(
            "/springflow/EmployeeEndpoint/create",
            {
                name: name,
                email: email,
                department: department,
            }
        );
    }

    static async findAll(): Promise<Employee[]> {
        return SpringFlowClient.get<Employee[]>(
            "/springflow/EmployeeEndpoint/findAll"
        );
    }

    static async findById(id: number): Promise<Employee> {
        return SpringFlowClient.get<Employee>(
            "/springflow/EmployeeEndpoint/findById",
            {
                id: id,
            }
        );
    }

    static async update(id: number, name: string, email: string, department: string, active: boolean): Promise<Employee> {
        return SpringFlowClient.get<Employee>(
            "/springflow/EmployeeEndpoint/update",
            {
                id: id,
                name: name,
                email: email,
                department: department,
                active: active,
            }
        );
    }

    static async delete(id: number): Promise<boolean> {
        return SpringFlowClient.get<boolean>(
            "/springflow/EmployeeEndpoint/delete",
            {
                id: id,
            }
        );
    }

}